refactor: unify crosschain token metadata description
